feat: Disable flash attention in Phi-3-vision-128k-instruct model
This commit disables the use of flash attention in the Phi-3-vision-128k-instruct model by setting the environment variable `USE_FLASH_ATTENTION_2` to `false`. This change improves the performance and efficiency of the model.

feat: Update torch version in test3.py

This commit updates the torch version in test3.py to the latest version.